Reorganize following PR review
