+++ app/src/main/java/com/example/dancognitionapp/AppViewModelProvider.kt
@@ -2,17 +2,15 @@ package com.example.dancognitionapp
 
 import android.app.Application
 import androidx.lifecycle.ViewModelProvider
-import androidx.lifecycle.createSavedStateHandle
 import androidx.lifecycle.viewmodel.CreationExtras
 import androidx.lifecycle.viewmodel.initializer
 import androidx.lifecycle.viewmodel.viewModelFactory
-import com.example.dancognitionapp.participants.edit.AddEditParticipantsFragment
 import com.example.dancognitionapp.participants.home.ParticipantsHomeViewModel
 import com.example.dancognitionapp.participants.edit.AddEditViewModel
 import com.example.dancognitionapp.participants.edit.ParticipantIdManager
 
 object AppViewModelProvider {
-    val Factory = viewModelFactory {
+    val factory = viewModelFactory {
         // Initializer for the ParticipantsViewModel
         initializer {
             ParticipantsHomeViewModel(
