+++ app/src/main/java/com/example/dancognitionapp/assessment/bart/ui/BartViewModel.kt
@@ -4,7 +4,6 @@ import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
 import com.example.dancognitionapp.assessment.bart.data.BalloonGenerator
-import com.example.dancognitionapp.assessment.bart.data.BartUiState
 import timber.log.Timber
 
 class BartViewModel: ViewModel() {
