+++ app/src/main/java/com/example/dancognitionapp/assessment/nback/data/NBackGenerator.kt
@@ -1,4 +1,4 @@
-package com.example.dancognitionapp.assessment.nback
+package com.example.dancognitionapp.assessment.nback.data
 
 import java.util.LinkedList
 import kotlin.random.Random
