+++ app/src/main/java/com/example/dancognitionapp/participants/edit/AddEditUiState.kt
@@ -1,5 +1,7 @@
 package com.example.dancognitionapp.participants.edit
 
+import com.example.dancognitionapp.participants.data.ParticipantDetails
+
 data class AddEditUiState(
     val participantDetails: ParticipantDetails = ParticipantDetails()
 )
